+++ CsharpExtras/Extensions/Helper/Dictionary/DictionaryComparisonImpl.cs
@@ -13,17 +13,22 @@ namespace CsharpExtras.Extensions.Helper.Dictionary
         /// <param name="thisCount">Number of elements in this dictionary</param>
         /// <param name="otherCount">Number of elements in other dictionary</param>
         /// <param name="firstMismatch">First mistmach, if one exists, or null if there are no mismatches</param>
-        public DictionaryComparisonImpl(int thisCount, int otherCount, (TKey, TVal)? firstMismatch)
+        /// <param name="otherValMistmatch">Other value for first mismatch, as a string, or null if other value not found</param>
+        public DictionaryComparisonImpl(int thisCount, int otherCount,
+            (TKey, TVal)? firstMismatch, string? otherValMistmatch)
         {
             ThisCount = thisCount;
             OtherCount = otherCount;
             FirstMismatch = firstMismatch;
+            OtherValMistmachOrNull = otherValMistmatch;
         }
 
         private int ThisCount { get; }
         private int OtherCount { get; }
 
         private (TKey, TVal)? FirstMismatch { get; }
+
+        private string? OtherValMistmachOrNull { get; }
 
         private (string, bool)? _messageAndIsEqual;
         private (string message, bool isEqual) MessageAndIsEqual => _messageAndIsEqual ??= GetMessageAndIsEqual();
@@ -37,7 +42,8 @@ namespace CsharpExtras.Extensions.Helper.Dictionary
             }
             if (FirstMismatch is (TKey k, TVal v))
             {
-                return ($"Mismatch found at key {k} and value {v}", false);
+                string otherVal = OtherValMistmachOrNull ?? "NOT FOUND";
+                return ($"Mismatch found at key {k}. This value: {v}. Other value: {otherVal}", false);
             }
             return ("Dictionaries are equal", true);
         }