+++ CsharpExtras/Map/Dictionary/Curry/CurryDictionaryRecursive.cs
@@ -42,7 +42,7 @@ namespace CsharpExtras.Map.Dictionary.Curry
 
         public CurryDictionaryRecursive(int arity, ICsharpExtrasApi api) : this((PositiveInteger)arity, api) { }
 
-        public CurryDictionaryRecursive(PositiveInteger arity, ICsharpExtrasApi api)
+        public CurryDictionaryRecursive(PositiveInteger arity, ICsharpExtrasApi api) : base(api)
         {
             Arity = (NonnegativeInteger)arity;
             _api = api;
@@ -130,7 +130,7 @@ namespace CsharpExtras.Map.Dictionary.Curry
             }
             else
             {
-                ICurryDictionary<TKey, TVal> curryChild = new NullaryCurryDictionary<TKey, TVal>(value);
+                ICurryDictionary<TKey, TVal> curryChild = new NullaryCurryDictionary<TKey, TVal>(value, _api);
                 AddDirectChild(firstKey, curryChild);
                 return true;
             }
@@ -174,15 +174,15 @@ namespace CsharpExtras.Map.Dictionary.Curry
             {
                 if (!other.ContainsKeyTuple(pair.keyTuple))
                 {
-                    return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, pair);
+                    return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, pair, null);
                 }
                 TVal otherValue = other.GetValueFromTuple(pair.keyTuple);
                 if (!isEqualValues(pair.val, otherValue))
                 {
-                    return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, pair);
+                    return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, pair, otherValue?.ToString());
                 }
             }
-            return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, null);
+            return new CurryDictionaryComparisonImpl<TKey, TVal>(Arity, otherArity, Count, otherCount, null, null);
         }
 
         private void AddDirectChild(TKey key, ICurryDictionary<TKey, TVal> curryChild)