+++ CsharpExtras/Map/Sparse/Compare/SparseArrayComparisonImpl.cs
@@ -13,14 +13,16 @@ namespace CsharpExtras.Map.Sparse.Compare
         /// <param name="thisUsedValuesCount">Number of elements in this array</param>
         /// <param name="otherUsedValuesCount">Number of elements in other array</param>
         /// <param name="firstMismatch">First mistmach, if one exists, including the keys and the underly array comparison</param>
+        /// <param name="otherValMistmatch">Other value for first mismatch, as a string, or null if other value not found</param>
         public SparseArrayComparisonImpl(int thisDimension, int otherDimension, int thisUsedValuesCount, int otherUsedValuesCount,
-            (IList<int> keyTuple, TVal val)? firstMismatch)
+            (IList<int> keyTuple, TVal val)? firstMismatch, string? otherValMistmatch)
         {
             ThisUsedValuesCount = thisUsedValuesCount;
             OtherUsedValuesCount = otherUsedValuesCount;
             ThisDimension = thisDimension;
             OtherDimension = otherDimension;
             FirstMismatch = firstMismatch;
+            OtherValMistmachOrNull = otherValMistmatch;
         }
 
         private int ThisDimension { get; }
@@ -29,6 +31,8 @@ namespace CsharpExtras.Map.Sparse.Compare
         private int OtherUsedValuesCount { get; }
 
         private (IList<int> keyTuple, TVal val)? FirstMismatch { get; }
+
+        private string? OtherValMistmachOrNull { get; }
 
         private (string, bool)? _messageAndIsEqual;
         private (string message, bool isEqual) MessageAndIsEqual => _messageAndIsEqual ??= GetMessageAndIsEqual();
@@ -45,7 +49,8 @@ namespace CsharpExtras.Map.Sparse.Compare
             }
             if (FirstMismatch is (IList<int> keyTuple, TVal val))
             {
-                return ($"Mismatch found at key tuple {string.Join(",", keyTuple)} and value {val}", false);
+                string otherVal = OtherValMistmachOrNull ?? "NOT FOUND";
+                return ($"Mismatch found at key tuple {string.Join(",", keyTuple)}. This value: {val}. Other value: {otherVal}", false);
             }
             return ("Arrays are equal", true);
         }