+++ CsharpExtras/Map/Sparse/TwoDimensional/SparseArray2DImpl.cs
@@ -12,19 +12,15 @@ namespace CsharpExtras.Map.Sparse.TwoDimensional
     {
         public ISparseArray<TVal> BackingArray { get; }
 
+        private NonnegativeInteger? _rowAxisIndex;
+        public NonnegativeInteger RowAxisIndex => _rowAxisIndex ??= (NonnegativeInteger)0;
+        private NonnegativeInteger? _columnAxisIndex;
+        public NonnegativeInteger ColumnAxisIndex => _columnAxisIndex ??= (NonnegativeInteger)1;
+
         public SparseArray2DImpl(ISparseArray<TVal> backingArray)
         {
             ValidateBackingArray(backingArray);
             BackingArray = backingArray ?? throw new ArgumentNullException(nameof(backingArray));
-        }
-
-        private void ValidateBackingArray(ISparseArray<TVal> backingArray)
-        {
-            PositiveInteger dimension = backingArray.Dimension;
-            if(dimension != 2)
-            {
-                throw new ArgumentException($"Dimension of backing array must be exactly 2. Found: {(int) dimension}");
-            }
         }
 
         public TVal this[int row, int column]
@@ -38,6 +34,12 @@ namespace CsharpExtras.Map.Sparse.TwoDimensional
             return BackingArray.CompareUsedValues(other.BackingArray, comparitor);
         }
 
+        public void InsertRows(int insertionIndex, int shiftVector) =>
+            BackingArray.Shift((NonnegativeInteger)0, insertionIndex, shiftVector);
+
+        public void InsertColumns(int insertionIndex, int shiftVector) =>
+            BackingArray.Shift((NonnegativeInteger)1, insertionIndex, shiftVector);
+
         public void SetArea(TVal[,] area, int leftmostRow, int topMostColumn)
         {
             ValidateArea(area, leftmostRow, topMostColumn);
@@ -48,26 +50,6 @@ namespace CsharpExtras.Map.Sparse.TwoDimensional
                     int rowAbs = leftmostRow + i;
                     int colAbs = topMostColumn + j;
                     BackingArray[rowAbs, colAbs] = area[i, j];
-                }
-            }
-        }
-
-        private void ValidateArea(TVal[,] area, int leftmostRow, int topMostColumn)
-        {
-            for (int i = 0; i < area.GetLength(0); i++)
-            {
-                int rowAbs = leftmostRow + i;
-                if (!BackingArray.IsValid(rowAbs, (NonnegativeInteger)0))
-                {
-                    throw new IndexOutOfRangeException($"Invalid row index: {i}");
-                }
-            }
-            for (int j = 0; j < area.GetLength(1); j++)
-            {
-                int colAbs = topMostColumn + j;
-                if (!BackingArray.IsValid(colAbs, (NonnegativeInteger)1))
-                {
-                    throw new IndexOutOfRangeException($"Invalid column index: {j}");
                 }
             }
         }
@@ -92,6 +74,35 @@ namespace CsharpExtras.Map.Sparse.TwoDimensional
                 }
             }
             return area;
+        }
+
+        private void ValidateArea(TVal[,] area, int leftmostRow, int topMostColumn)
+        {
+            for (int i = 0; i < area.GetLength(0); i++)
+            {
+                int rowAbs = leftmostRow + i;
+                if (!BackingArray.IsValid(rowAbs, RowAxisIndex))
+                {
+                    throw new IndexOutOfRangeException($"Invalid row index: {i}");
+                }
+            }
+            for (int j = 0; j < area.GetLength(1); j++)
+            {
+                int colAbs = topMostColumn + j;
+                if (!BackingArray.IsValid(colAbs, ColumnAxisIndex))
+                {
+                    throw new IndexOutOfRangeException($"Invalid column index: {j}");
+                }
+            }
+        }
+
+        private void ValidateBackingArray(ISparseArray<TVal> backingArray)
+        {
+            PositiveInteger dimension = backingArray.Dimension;
+            if (dimension != 2)
+            {
+                throw new ArgumentException($"Dimension of backing array must be exactly 2. Found: {(int)dimension}");
+            }
         }
     }
 }